LayoutTests/http/wpt/webauthn/public-key-credential-get-success-local.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with minimum options in a mock local authenticator.
PASS PublicKeyCredential's [[get]] with matched allow credentials in a mock local authenticator.
